Workaround for publishing SBT plugins using maven style
When the artifact is an SBT plugin (`sbtPlugin := true`) and is
configured to be published using maven style
(`publishMavenStyle := true`) instead of stripping the artifact
minor scala version and minor sbt version it uses the content
bintray API to push the artifacts.

Bintray does not allow pushing artifacts using the maven upload
API as SBT plugin layout is incompatible with the maven one.

The previous behaviour is kept when the project is not an SBT
plugin or the setting to publish using the maven style is set to
false.

Related to #125, sbt/sbt#3410
Fixes #72