Ensure that we don't needlessly create HTTP clients and leak threads
ConcurrentMap.getOrElseUpdate may run the thunk without eventually
insering the computed value in the map, if multiple threads are
calling it for a given key. This situation seems to happen in SBT
setting evaluation, perhaps in multi-module builds. The upshot is
that Netty creates 2 x $numcpus "New NIO Worker" threads that
live until the JVM exits.

This commit adds synchronization around the interaction with the
cache to ensure a given cache key only creates a single HTTP client
and thread pool.

Future work could reuse a HTTP client for all of Bintray's usage.
I haven't made that change because I can't find documentation of
the threadsafety of dispatch.Http.